♻️ user collec > refacto split sub comps
