+++ src/containers/User/Collection/Gentks.module.scss
@@ -1,0 +1,5 @@
+.select_comp_container {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+}
