+++ src/containers/User/UserCollection.module.scss
@@ -5,12 +5,6 @@
   justify-content: flex-end;
 }
 
-.select_comp_container {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
-
 .article_container {
   margin-top: $spacing-3x-large;
   width: min(100%, 1200px);
